+++ jax/_src/numpy/lax_numpy.py
@@ -2362,15 +2362,7 @@ def _check_forgot_shape_tuple(name, shape, dtype) -> str | None:  # type: ignore
 
 @util.implements(np.array_equal)
 def array_equal(a1: ArrayLike, a2: ArrayLike, equal_nan: bool = False) -> Array:
-  try:
-    a1, a2 = asarray(a1), asarray(a2)
-  except Exception as err:
-    # TODO(jakevdp): Deprecated 2023-11-23; change to error.
-    warnings.warn("Inputs to array_equal() cannot be coerced to array. "
-                  "Returning False; in the future this will raise an exception.\n"
-                  f"{err!r}",
-                  DeprecationWarning, stacklevel=2)
-    return bool_(False)
+  a1, a2 = asarray(a1), asarray(a2)
   if shape(a1) != shape(a2):
     return bool_(False)
   eq = asarray(a1 == a2)
@@ -2381,15 +2373,7 @@ def array_equal(a1: ArrayLike, a2: ArrayLike, equal_nan: bool = False) -> Array:
 
 @util.implements(np.array_equiv)
 def array_equiv(a1: ArrayLike, a2: ArrayLike) -> Array:
-  try:
-    a1, a2 = asarray(a1), asarray(a2)
-  except Exception as err:
-    # TODO(jakevdp): Deprecated 2023-11-23; change to error.
-    warnings.warn("Inputs to array_equiv() cannot be coerced to array. "
-                  "Returning False; in the future this will raise an exception.\n"
-                  f"{err!r}",
-                  DeprecationWarning, stacklevel=2)
-    return bool_(False)
+  a1, a2 = asarray(a1), asarray(a2)
   try:
     eq = ufuncs.equal(a1, a2)
   except ValueError: